pvr: Fix hybrid mode signaling TA completion too early
The dma_next_list() function was setting the "list transferred" flag for
the lists that used DMA to transfer the vertices and also for the other
ones. It also signaled completion through the semaphore after the last
transfer completed.

The problem with that, is that there is a difference between "the vertex
data has been transferred to the TA" and "the TA is done processing the
lists"; and the "list transferred" flags are to be set in the second
case.

Address this issue by leaving it to the IRQ handler to set the "list
transferred" flags and handle the completion semaphore.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>