pvr: Fix hang when enabled lists aren't submitted
When two lists were enabled, one for direct-rendering and one for
deferred (DMA) rendering, but the first one was not used, only the IRQ
corresponding to the second list fired, causing the IRQ handler to not
start the PVR CORE rendering step, as it waited for all enabled lists to
be transferred.

Address this issue by adding dummy vertices to enabled lists that were
not used in the current scene.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>