Add support for Store Queues with MMU (#779)
* mmu: Add function mmu_enabled()

This function can be used to know whether or not the MMU is in use.

* mmu: Support address translation of store queues

Reserve the last two TLB entries and map a 1 MiB page each for use
with the SQ area.

Add the function mmu_set_sq_addr(), which will overwrite the physical
address that the SQ area points to in the TLB.

* sq: Update sq_lock() prototype to work with MMU

Change its prototype: it now returns a pointer, which is the translated
address of the source pointer. In non-MMU use, it's returns the same
value as SQ_MASK_DEST(src).

When MMU is enabled, sq_lock() will reset the address translation for
the SQ memory area, and convert the source address as an offset to that
memory area, making it possible to use SQs when MMU is enabled.

The caveat is that we only map one megabyte of SQ memory area for
simplicity, and therefore the SQ memory window is reduced.
This should be fine though, as apart from e.g. a whole-VRAM clear,
nobody will ever need such a big window.

* sq: Update sq_cpy() to work with MMU

When using the MMU the SQ area is a 1 MiB page, so we must not copy data
outside this area.

* sq: Update sq_set32() to work with MMU

When using the MMU the SQ area is a 1 MiB page, so we must not copy data
outside this area.

* mmu: Remove mmu_disable() / mmu_restore()

Now that SQs work with MMU, those two functions aren't useful anymore
and can be dropped.

---------

Signed-off-by: Paul Cercueil <paul@crapouillou.net>