pvr: Add pvr_init() flag to disable vertex buffer double-buffering
Having a double-buffering mechanism for the vertex buffer makes it
possible to upload geometry data to the Tile Accelerator while the PVR
is busy rendering the previous scene.

However, it means having two buffers for the geometry data. Some
applications may prefer using a much smaller vertex buffer to lower
VRAM usage, even if it means the TA and PVR cannot be used at the same
time.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>