Drop vec3f ASM
Use the vec3f_t type defined by the header, and replace all inline ASM
with C code.

Apart from the "fipr" instruction, which we already have macros for,
there is absolutely no reason to have these algorithms in ASM. Not only
it is unnecessary, it is also harmful as the C compiler cannot
re-organize those instructions as it sees fit.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>