spinlock: Update inline ASM
Instead of specifying that the inline ASM modifies memory without more
information, tell it that the memory pointed by the spinlock pointer
will be modified.

The compiler then knows that it does not have to reload all the
variables from memory after a call to spinlock_trylock() or
spinlock_lock().

Signed-off-by: Paul Cercueil <paul@crapouillou.net>