pvr: Do not check DMA completion in pvr_sq functions
The only reason why you would want to wait for any pending DMA transfer
to complete before using the SQs is when the DMA transfer targets the
TA, in which case a SQ transfer to the TA would be asking for trouble.

However, the PVR scene code is already making sure that the TA is ready
before a new list can be opened; this means that it is guaranteed that
lists configured for DMA will be done transferring before anything is
transferred to the TA via SQs.

Therefore, with the current driver code, there is no point in waiting
for DMA completion.

With that said, the code actually did not wait for DMA completion, but
was doing something much worse: in the case where a DMA transfer was
in progress, it would simply display an error, and return without
processing the SQ transfer. This caused heavy errors, including crashes,
for instance when a list's "EOL marker" could not be added.

Address this issue by simply not checking for the DMA completion in the
pvr_sq functions.

Fixes #878.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>