Remove dependence on minimal termios stub.
For #636 a stub of termios functionality was added
in order to support isatty. This broke libjimtcl and
would break other software that detected termios
compatibility as it implemented only the minimum
needed for isatty.

This replaces the use of tcgetattr with a check of
the kos vfs handler to test if it's pty. The stub
termios header and tcgetattr is then removed.