cache: Improve dcache_alloc_block()
Instead of using the "memory" clobber, instruct the compiler that the 32
bytes located at the address passed as parameter have been written.

The compiler then knows that anything read or written to that memory area
prior to the dcache_alloc_block() call is dead code, and anything read or
written outside this area is still valid after the call.

The "volatile" specifier has been dropped as well, given that the
compiler has now all the information it needs, and can freely reorganize
the instructions as needed.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>