environ.sh: Enable fast-math by default
We recommend -ffast-math everywhere, and everyone seems to be using it.
Those who don't, probably don't need the full IEEE compliance and just
forgot to add the flag.

Note that the previous comment suggested that FSQRT was not used without
fast-math, but this was wrong. GCC will still emit that instruction in a
"fast path" when it knows it can use it, or a call to the C sqrtf()
otherwise.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>