Remove hack of detecting a dir as fd > 100
It seems this awful hack was possible due to dcload returning
the host-side `DIR *` as the fd for dirs. The problem is that
it's absolute bunk as there's no guarantee that the fds for
non-dirs from the host side returned by `open` will be less
than 100. This could mean that a file gets opened and not be
able to be closed properly. This also clears the way to remove
the hack from the host-side dc-loads, and to correct readdir
behavior.