perfcntr: Handle counter overflow while reading values
Read the high part twice, before and after reading the high part,
to make sure that the low counter didn't overflow. We can detect an
overflow by the two reads of the high part returning different
values.

Note that we don't mask the low 16 bits anymore, as the upper 16 bits
are zero anyway.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>