perfcntr: Slightly reduce latency of perf_cntr_timer_ns()
To reduce the latency of perf_cntr_timer_ns(), perform the
perf_cntr_timer_enabled() check *before* reading the counter.

The argument that the counter should be read first as to not "record
overhead" is bogus; the reason being, if you use these timers to
measure the time taken between points A and B, you will always end up
measuring the "perf_cntr_timer_enabled()" overhead, just at a different
extremity.

Therefore, if we "record overhead" in both cases, it's better to use the
case where we have slightly lower latency.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>