pvr: Flush source buffer in pvr_dma_transfer()
The pvr_dma_transfer(), as well as the pvr_txr_load_dma(),
pvr_dma_load_ta() and pvr_dma_yuv_conv() functions, expected the source
data to be completely flushed to RAM, without it being explicitly
written anywhere. Therefore, the expectation was that manual handling
of the cache was not required.

Update the pvr_dma_transfer() function (called by the other ones) to
handle the cache flush itself.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>