Added thread attribs to cont_btn_callback worker.
1) When using the cont_btn_callback() mechanism, a new thread is
   created, which is left unnamed. This looks kind of janky and
confusing when seen in the thread listing when dumping thread info or
after an exception.
    - Added a default label to the worker thread to identify it
2) By far, the most commonly implemented callback with this mechanism
   tends to be some absolutely trivial logic like a simple call to
exit(0). Because of this, games which are low on RAM are typically
paying an exuberantly high amount of memory overhead just to call one
function with a default stack size of 32KB for the worker.
    - Added a default stack size of 8KB, which can be overridden at
      compile-time, consistently with how other default thread sizes can
be overwritten.