pvr: Fix dma_next_list() racing with itself
The very first call to dma_next_list() does not run in an interrupt
context, nor does it run with the IRQs disabled.

As soon as pvr_dma_load_ta() is called, it is therefore possible to have
dma_next_list() re-entered from within the DMA complete interrupt
handler. This is fine, as long as we make sure that the function does
not do anything after this point that could race with the interrupt
handler.

One possible race was with the DMA complete interrupt firing before the
"pvr_state.lists_dmaed" was updated, in which case the interrupt handler
wouldn't get the updated value, or when the interrupt was firing in the
middle of updating the "pvr_state.lists_dmaed".

Signed-off-by: Paul Cercueil <paul@crapouillou.net>