Ensure sq_writeback is flagged as causing memory to change.
First reported with #897, a regression from #877. While it's
appropriate for the generalized dcache_pref_block to not have
outputs, setting the same for the store queue writeback causes
it to be optimized away in some situations (specifically DR).

This simply rolls back to the previous behavior for the sq case.
It should be possible to optimize further, but different attempts
led to build failures.