cache: Fix non-working dcache_{flush,purge}_all()
The algorithm of dcache_flush_all() and dcache_purge_all() were
sometimes exhibiting problems under unidentified conditions, in which
case some cache lines were not written back properly to RAM.

This caused all kind of memory corruption problems, when the PVR DMA was
used for transfers with a size above the threshold at which the
"dcache_flush_all()" function is called.

Fix these functions by accessing the cache address array using
association. Since the underlying problem is not fully understood, the
fix is not fully understood either, but it was reported as fixing the
memory corruption in Doom64.

Fixes #928.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>