Allow patched toolchain to support -m4-nofpu
This still will not build language support for objc/cpp but does
for c/cpp and can successfully build dcload with a patch there.
This does not support building KOS but as startup.s is provided
to the toolchain, it needed adjustment.

This should be helpful for those who had been investigating using
KallistiOS org stuff on alternative SH machines.