VMU Button Bitflag Fix, Plus Maintaining Previous Button States. (#915)
* Fixed #defines for VMU button bitflags.

- Bitflags were never correctly defined for the VMU buttons in vmu.h,
  but it was never caught, since you can also access them via bitfields.
- Corrected the incorrect flags.

* Added tracking for previous VMU button state.

In order to do anything useful in terms of responding to button state
change events, such as when a button is FIRST pressed or FIRST released,
any app would have to maintain the previous state of the buttons in
order to compare them to the current state.

We've been moving this functionality into KOS, so I've done the same
with the VMU buttons. Since the maple state data is static for every
device, we aren't storing any extra information anyway. Just using one
more byte of the 1024 byte buffer at the end of our device (for a
whopping 2 bytes of total button info).

- Pulled VMU buttons out of state and into their own union of raw button
  masks and individual bitfields.
- Added logic to driver to cache the previous frame's state
- Created new vmu_state_t struct exposing these two frames' worth of
  button state info.

* Made vmu_cond_t struct private.

This struct is only for processing the raw maple response packet when
querying for button states... There was no reason to expose it as
something cluttering up the main API and confusing our users.

- Moved vmu_cond_t from vmu.h to vmu.c.
- Updated copyright information.