Remove redefined 'BSD-style' types.
This should help reduce the possible impact of #840
which describes how types defined in this file escape
out and can conflict with other codebases.

This is in-line with our general move to stdint types
and, as far as I can tell, there was only one use of
these types in the codebase.