Remove perfcounter usage from time tracking functions.
There were a number of issues with the usage of perf counters here.
The biggest was that it violates the arch interface into KOS by
directly digging into a dreamcast-specific driver from the outer
threading/libc support. The second is that it simply unnecessary.
I don't see anything in the descriptions of the usage to indicate
that this has merits over the standard ns timer used elsewhere, and
the mere usage of a separate timer with different qualities means
that it's more likely to have discrepancies with our general ns
timer-backed things.
The third is that it's not accurate. Due to not counting sleep time
the performance counters horribly undercount the amount of time spent
in the idle thread, so the total time percentage of a `thd_pslist`
would rarely add up to 100%.
The last is that doing so allows us to move forward with a better
way to streamline a perfcounter based timer into the API.