Ensure precision and consistency of the thread list printout.
A number of changes were made in order to ensure the total
timing of the thread list cpu_times would always reach 100%.
First, IRQs are disabled before taking the first time measurements.
This is important as, without doing so we could context switch away
and skew the percentages. This was happening randomly in the general
threading test.

Second, we get the total time `ns_time` only once at the start of the
listing. Otherwise when checking it repeatedly it will naturally skew
as it grows duing the process of printing. To match with this, we update
the current thread's time just before reading it. The small amount of time
between those two operations ends up factoring into the new [system]
line item.

Third, adding a new line item [system]. This tracks the difference between
the total `cpu_time` recorded for each thread and the clock. This not only
helps account for the small chunk of time mentioned before but all the time
taken between the updating of one thread's time and the new schedule time
of the next thread that starts. With basic operation should should mostly
track just context switching time, but can be significant when we're doing
something like genwaiting on multiple objects as we'll be repeatedly
jumping out of the idle thread to update them.

With these changes the total percentages should always add up to exactly
100% (permitting possible rounding error) while before they would vary
wildly in situations with lots of waits and idle time.