fs_vmu: Fix seeking in files with headers
My recent change with VMU headers worked as long as applications did not
try to use fs_seek(). The problem was then that the beginning of the
file was set relative to the actual beginning, before the header, and
not after the header.

The fix is to consider that the cursor position is always relative to
the start of the file, and the start of the file being the address after
the file header, if there is one.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>