Add kbd_set_repeat_timing, clean up repeating internals.
The two loose static vars have been updated into a struct,
and an accessor added to set the values. Additionally a
small change was made so that if the repeat timing is set to 0
the timing is not pulled or tested at all. Previously, each
poll would unnecessarily check the timer and store/test times.