Un-hide code behind ifdef guards
Hiding code behind ifdef guards has the nasty issue that the code is
only seen by the compiler under a specific configuration. As time
passes, the guarded code might become stale, causing some rarely used
configurations to fail to build.

This can be solved by using the __is_defined() macro to conditionally
execute the code only if the given guard macro is defined. The
difference is that the compiler will see the code, and report issues if
it cannot build it, but ultimately drop it as dead code if the guard
macro is not defined.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>