irq: Don't deref fp twice.
In #600 there was a change to store `arch_fptr_ret_addr` as
it was being tested as valid before being printed. Unfortunately
this was done in a way that overwrote the current fp being worked
on. So when it came time to pass it to `arch_fptr_next` another
layer of deference was happening.