cdrom: Use cdrom_poll() in cdrom_exec_cmd_timed()
Factorize code by using cdrom_poll() in cdrom_exec_cmd_timed().

This commit also fixes cdrom_exec_cmd_timed() being extremely slow;
that's because after the command was sent to the GD drive, the response
was only obtained after the next VBLANK, which meant having to wait for
up to 16ms.

Now, instead of waiting on the "cmd_done" semaphore signaled in the
vblank interrupt, we poll for the command response until we obtain it.
It will cause a slight CPU overhead but will be much faster.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>