cdrom: Simplify vblank code
Since nothing waited on the "cmd_done" semaphore anymore, the code
signaling it as well as the semaphore itself can be dropped.

The timeout code can also be dropped, as it was now handled by the
cdrom_poll() function.

The VBLANK interrupt handler now only has to check for errors when DMA
is used.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>