examples: replace `.keepme` with `.gitignore`
In each case that we were using `.keepme`, we had an empty romdisk
folder that is either populated by generated files, or in the case
of the sound examples user provided files. Instead of having a
`.gitignore` in the main folder of the example that lists the files
to ignore in the romdisk, then a separate `.keepme` just to keep the
dir alive, we can have the `.gitignore` be directly in the romdisk
folder. To match this, the Makefile.rules was updated to have romdisk
building ignore `.gitignore` (which seems like just a nice thing
anyways for users).

Additionally, the sound examples previously did not have any `.gitignore`
so so it was added to prevent the `test.mp3/ogg/opus/adx` files from
appearing as untracked files.