pvr: Limit texture addresses to VRAM size
Previously, the various compile functions would apply a 0x00fffff8 mask
to the texture address. This is strange as it limits the address space
to 16 MiB while the Dreamcast has only 8 MiB of video RAM.

Use a new mask of (PVR_RAM_SIZE - 1) to address this issue, and make it
more obvious what the mask is for.

Note that the low 3 bits don't need to be cleared as the address is
shifted right by 3 later.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>