dbglog: Add option to compile-time disable/limit debug logging.
This should help allow debug messages to be garbage collected,
and provide simpler steps for a release build when paired with
NDEBUG. `DBGLOG_LEVEL_SUPPORT` can be set to a dbglog level in
order to cap the level available at runtime to that amount.
`DBGLOG_DISABLED` overrides this down to disallow any output.

As the actual function backing `dbglog` is now named differently
this had to be updated in the exports. I've additionally moved it
out of the stdio section to reflect that it's kos-internal and
no longer is being shared with stdio after #1000 .