dc-chain: Remove fake KOS objects from libgcc after build
This should not be necessary and I don't understand why we need this;
the symbols inside fake-kos.o are all weak, and should all be replaced
by symbols provided by libkallisti.a. Alas, this does not seem to work
reliably, and cause crashes and issues that are hard to debug.

Address these issues by just dropping the fake-kos.o object from the
libgcc.a archive.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>