examples: More stable exec example
The changes made in #1112 to make the romdisk vfs
dynamically allocate file handles had a side effect
of being more likely to break this exec example.
This happens because the example was using `fs_mmap`
with a hidden assumption that the memory allocated for
the open rd file would persist through the
`arch_shutdown` that happens during `arch_exec`.

This only ever worked because of the specifics of how the
example were being crafted: having the subelf in a static
romdisk built into the loaded parent binary and could have
failed before #1112 with any mildly different use like loading
a romdisk from a disc first.