vmu_lcd example: Add wait after draw to throttle animation.
The example never attempted to throttle itself because the
screen drawing commands were synchronous and would wait to
receive a response from the VMU before moving forward with
execution. With the changes in #904 there is no wait and
the user is responsible for throttling the drawing.

Without this the example is illegible on the VMU screen,
updating faster than can be seen. This wait amount gets it
back down to a pace similar to before.