Compile-test gthr-kos.h with GCC >= 15
Starting from GCC 15, the C23 standard is supported.

This is important, because this standard allows the redefinition of
types, as long as the new definition matches the type of the old
definition.

Therefore, we can compile the gthr-kos.h header with the
TEST_GTHR_KOS_API macro set, which will cause all the headers that
define the symbols used in there to be included. The effect of that, is
that all the re-defined functions and types in gthr-kos.h will be
compared vs. the original one in KOS headers, and GCC will throw an
error if any of them has a different type or prototype.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>